# Assignment 9/Workforce-management/src/WorkforceManagementAPI.WEB/openapi.json
openapi: 3.0.1
info:
  title: WorkforceManagementAPI.WEB
  version: v1
paths:
  /api/Team:
    get:
      tags:
        - Team
      summary: List all teams, existing in the database.
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
          
    post:
      tags:
        - Team
      summary: Create a team.
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
  /api/Team/{teamId}:
    get:
      tags:
        - Team
      summary: Find a team by its Id.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
    put:
      tags:
        - Team
      summary: Edit a team.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/TeamRequestDTO'
      responses:
        '200':
          description: OK - Request succeeded.
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponseDTO'
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
    delete:
      tags:
        - Team
      summary: REMOVE a team.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
  /api/Team/MyTeams:
    get:
      tags:
        - Team
      summary: List all teams, the logged user is member of.
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
  /api/Team/{teamId}/Assign/{userId}:
    post:
      tags:
        - Team
      summary: Assign user to a team.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
        - name: userId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            nullable: true
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
  /api/Team/{teamId}/Unassign/{userId}:
    delete:
      tags:
        - Team
      summary: REMOVE user from a team.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
        - name: userId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            nullable: true
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
  /api/Team/{teamId}/AssignLeader/{userId}:
    put:
      tags:
        - Team
      summary: Assign user as a TEAM LEADER.
      parameters:
        - name: teamId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            format: uuid
        - name: userId
          in: path
          description: ''
          required: true
          schema:
            type: string
            description: ''
            nullable: true
      responses:
        '200':
          description: OK - Request succeeded.
        '201':
          description: Created - Request resulted in new resource created.
        '401':
          description: Unauthorized - Please check the provided credentials.
        '403':
          description: Forbidden - Your credentials don't meet the required authorization level to access the resource. Please, contact your administrator to get desired permissions.
        '404':
          description: NotFound - Requested information does not exist in the server.
        '409':
          description: Conflict - The submitted entity ran into a conflict with an existing one.
        '400' :
          description: BadRequest - Request could not be understood by the server.
        '500' :
          description: InternalServerError - Generic error occured in the server.
...
components:
  schemas:
    TeamResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creatorId:
          type: string
          nullable: true
        modifierId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date
        modifiedAt:
          type: string
          format: date
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        teamLeaderId:
          type: string
          nullable: true
      additionalProperties: false
...
    TeamRequestDTO:
      required:
        - description
        - title
      type: object
      properties:
        title:
          maxLength: 50
          type: string
        description:
          maxLength: 200
          type: string
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization: Bearer {token}"'
      name: Authorization
      in: header
security:
  - Bearer: []